addFilm.jsp 작성 LanguageMapper 추가 LanguageService 추가 addActor or addActorFile 파일 첨부 시 이미지 파일만 등록 하도록 분기 코드 추가

+++ src/main/java/com/example/sakila/vo/Film.java
@@ -12,9 +12,9 @@ public class Film {
 	private int languageId; // FK → language.language_id
 	private int originalLanguageId; // FK → language.language_id
 	private int rentalDuration;
-	private int rentalRate;
+	private double rentalRate;
 	private int length;
-	private int replacementCost;
+	private double replacementCost;
 	private String rating;
 	private String specialFeatures;
 	private String lastUpdate;
